zsh: unset nomatch
If a pattern for filename generation has no matches, don't print an error,
instead leave it unchanged in the argument list.

http://zsh.sourceforge.net/Doc/Release/Options.html

With the kind of globs zsh supports, we frequently hit an error for rake,
arguments,  ^-regexes, and the like. In these cases I want them passed through
as-is to the underlying programs.