git: don't make merge commits in merge-pr
I originally had this for maybe_haskell, where it was important that I
signed all commits to the book, and so needed the merge commit for any
outside contributions. In general, I'd rather maintain linear history
than ensure contributor commits are signed.

There may be a better way to do this.