Add pre-up check that $RCRC is configured
On a new machine, forgetting to install the correct `.rcrc`
independently will result in an `rcup` that spams my system with
incorrect dotfiles. This can be annoying to undo, since a subsequent
`rcdn` may be affected by the now-installed `.rcrc`.

Note that this will not currently stop the `rcup` (see this [issue][]),
but in the meantime just seeing the message will be useful still.

[issue]: https://github.com/thoughtbot/rcm/issues/288