[git] Fix mergetool
I always wanted to set up the middle pane on the bottom, which you get
if you run `^w-l`, `^w-J` after opening it. But this `$wincmd` always
sent the furthest left pane to the bottom. Weird.

Well, I finally figured out that `:wincmd` is different than `$wincmd`
in this regard and works correctly. Silly.